style(main): norm done
