+++ src/graphic/collision.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   collision.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pdal-mol <pdal-mol@student.42.fr>          +#+  +:+       +#+        */
+/*   By: aandric <aandric@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/26 13:00:31 by pdal-mol          #+#    #+#             */
-/*   Updated: 2022/11/03 13:23:14 by pdal-mol         ###   ########.fr       */
+/*   Updated: 2022/11/03 13:43:45 by aandric          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -116,24 +116,12 @@ t_ray	*get_collision_coord(t_data *data)
 {
 	t_ray	*ray_horizontal;
 	t_ray	*ray_vertical;
-	t_bool	is_horizontal;
 
 	ray_vertical = get_collision_x(data);
 	ray_horizontal = get_collision_y(data);
-	is_horizontal = false;
-	if (!ray_vertical->coll || !ray_horizontal->coll)
-	{
-		if (!ray_vertical->coll)
-			is_horizontal = true;
-	}
-	else
-	{
-		ray_vertical->len = get_ray_len(data, ray_vertical);
-		ray_horizontal->len = get_ray_len(data, ray_horizontal);
-		if (ray_horizontal->len < ray_vertical->len)
-			is_horizontal = true;
-	}
-	if (is_horizontal)
+	ray_vertical->len = get_ray_len(data, ray_vertical);
+	ray_horizontal->len = get_ray_len(data, ray_horizontal);
+	if (ray_horizontal->len < ray_vertical->len)
 	{
 		free(ray_vertical);
 		return (ray_horizontal);
