+++ src/graphic/colors.c
@@ -6,7 +6,7 @@
 /*   By: aandric <aandric@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/03 11:49:37 by aandric           #+#    #+#             */
-/*   Updated: 2022/11/03 13:24:14 by aandric          ###   ########lyon.fr   */
+/*   Updated: 2022/11/03 13:33:24 by aandric          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,7 @@
 
 int	rgb_to_hex(unsigned char **rgb)
 {
-	return ((ft_atoi((const char *)rgb[0]) & 0xff) << 16) + (ft_atoi((const char *)rgb[1]) & 0xff << 8) + (ft_atoi((const char *)rgb[2]) & 0xff);
+	return ((ft_atoi((const char *)rgb[0]) & 0xff) << 16) 
+		+ (ft_atoi((const char *)rgb[1]) & 0xff << 8)
+		+ (ft_atoi((const char *)rgb[2]) & 0xff);
 }
