+++ src/parsing/get_features_from_scene.c
@@ -6,16 +6,27 @@
 /*   By: aandric <aandric@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/13 16:25:27 by pdal-mol          #+#    #+#             */
-/*   Updated: 2022/11/03 13:31:39 by aandric          ###   ########lyon.fr   */
+/*   Updated: 2022/11/03 13:50:33 by aandric          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/cub3d.h"
 
+char	*remove_key(char *line, int key_end)
+{
+	char	*buff;
+
+	if (ft_strlen(&line[key_end]) == 0)
+		return (NULL);
+	buff = ft_strdup(&line[key_end]);
+	if (!buff)
+		error(MEMALLOC);
+	return (buff);
+}
+
 char	*get_line_from_key(char **parsed_scene, char *key)
 {
 	int		i;
-	char	*buff;
 
 	i = 0;
 	while (parsed_scene[i])
@@ -23,20 +34,8 @@ char	*get_line_from_key(char **parsed_scene, char *key)
 		if (ft_strncmp(parsed_scene[i], key, ft_strlen(key)) == 0)
 		{
 			if (parsed_scene[i][0] == 'C' || parsed_scene[i][0] == 'F')
-			{
-				if (ft_strlen(&parsed_scene[i][1]) == 0)
-					return (NULL);
-				buff = ft_strdup(&parsed_scene[i][1]);
-				if (!buff)
-					error(MEMALLOC);
-				return (buff);
-			}
-			if (ft_strlen(&parsed_scene[i][2]) == 0)
-				return (NULL);
-			buff = ft_strdup(&parsed_scene[i][2]);
-			if (!buff)
-				error(MEMALLOC);
-			return (buff);
+				return (remove_key(parsed_scene[i], 1));
+			return (remove_key(parsed_scene[i], 2));
 		}
 		i++;
 	}
